requirements.txt
python-dotenv==0.20.0
selenium==4.2.0
webdriver-manager==3.7.0
PyYAML==6.0